+++ core/bbs.go
@@ -1,5 +1,0 @@
-package core
-
-func (app *Application) bbsGetTest() string {
-	return "BBs - test"
-}