+++ core/interfaces.go
@@ -1,44 +1,28 @@
 package core
 
+import (
+	"core-building-block/core/model"
+
+	log "github.com/rokmetro/logging-library/loglib"
+)
+
 //Services exposes APIs for the driver adapters
 type Services interface {
-	SerGetVersion() string
-	SerGetAuthTest() string
-	SerGetCommonTest() string
-}
-
-type servicesImpl struct {
-	app *Application
-}
-
-func (s *servicesImpl) SerGetVersion() string {
-	return s.app.serGetVersion()
-}
-
-func (s *servicesImpl) SerGetAuthTest() string {
-	return s.app.serGetAuthTest()
-}
-
-func (s *servicesImpl) SerGetCommonTest() string {
-	return s.app.serGetCommonTest()
+	SerGetVersion(l *log.Log) string
+	SerGetAuthTest(l *log.Log) string
+	SerGetCommonTest(l *log.Log) string
 }
 
 //Administration exposes administration APIs for the driver adapters
 type Administration interface {
 	AdmGetTest() string
 	AdmGetTestModel() string
-}
 
-type administrationImpl struct {
-	app *Application
-}
+	AdmCreateGlobalConfig(setting string) (*model.GlobalConfig, error)
+	AdmGetGlobalConfig() (*model.GlobalConfig, error)
+	AdmUpdateGlobalConfig(setting string) error
 
-func (s *administrationImpl) AdmGetTest() string {
-	return s.app.admGetTest()
-}
-
-func (s *administrationImpl) AdmGetTestModel() string {
-	return s.app.admGetTestModel()
+	AdmCreateOrganization(name string, requestType string, requiresOwnLogin bool, loginTypes []string, organizationDomains []string) (*model.Organization, error)
 }
 
 //Encryption exposes APIs for the Encryption building block
@@ -46,30 +30,20 @@ type Encryption interface {
 	EncGetTest() string
 }
 
-type encryptionImpl struct {
-	app *Application
-}
-
-func (s *encryptionImpl) EncGetTest() string {
-	return s.app.encGetTest()
-}
-
 //BBs exposes users related APIs used by the platform building blocks
 type BBs interface {
 	BBsGetTest() string
 }
 
-type bbsImpl struct {
-	app *Application
-}
-
-func (s *bbsImpl) BBsGetTest() string {
-	return s.app.bbsGetTest()
-}
-
 //Storage is used by core to storage data - DB storage adapter, file storage adapter etc
 type Storage interface {
 	SetStorageListener(storageListener StorageListener)
+
+	CreateGlobalConfig(setting string) (*model.GlobalConfig, error)
+	GetGlobalConfig() (*model.GlobalConfig, error)
+	SaveGlobalConfig(setting *model.GlobalConfig) error
+
+	CreateOrganization(name string, requestType string, requiresOwnLogin bool, loginTypes []string, organizationDomains []string) (*model.Organization, error)
 }
 
 //StorageListener listenes for change data storage events
@@ -78,7 +52,7 @@ type StorageListener interface {
 }
 
 type storageListenerImpl struct {
-	app *Application
+	app *application
 }
 
 func (a *storageListenerImpl) OnAuthConfigUpdated() {