+++ driven/storage/adapter.go
@@ -1,21 +1,47 @@
 package storage
 
 import (
+	"context"
 	"core-building-block/core"
 	"core-building-block/core/auth"
+	"core-building-block/core/model"
 	"errors"
 	"fmt"
-	"log"
 	"strconv"
 	"time"
 
+	"github.com/google/uuid"
+	log "github.com/rokmetro/logging-library/loglib"
 	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/bson/primitive"
+	"go.mongodb.org/mongo-driver/mongo"
 )
+
+type organization struct {
+	ID               string   `bson:"_id"`
+	Name             string   `bson:"name"`
+	Type             string   `bson:"type"`
+	RequiresOwnLogin bool     `bson:"requires_own_login"`
+	LoginTypes       []string `bson:"login_types"`
+
+	Config organizationConfig `bson:"config"`
+
+	DateCreated time.Time  `bson:"date_created"`
+	DateUpdated *time.Time `bson:"date_updated"`
+}
+
+type organizationConfig struct {
+	ID      string   `bson:"id"`
+	Domains []string `bson:"domains"`
+
+	DateCreated time.Time  `bson:"date_created"`
+	DateUpdated *time.Time `bson:"date_updated"`
+}
 
 //Adapter implements the Storage interface
 type Adapter struct {
-	db *database
+	db     *database
+	logger *log.StandardLogger
 }
 
 //Start starts the storage
@@ -69,15 +95,112 @@ func (sa *Adapter) LoadAuthConfigs() (map[string]auth.AuthConfig, error) {
 	return authInfoMap, nil
 }
 
+//CreateGlobalConfig creates global config
+func (sa *Adapter) CreateGlobalConfig(setting string) (*model.GlobalConfig, error) {
+	globalConfig := model.GlobalConfig{Setting: setting}
+	_, err := sa.db.globalConfig.InsertOne(globalConfig)
+	if err != nil {
+		return nil, err
+	}
+	return &globalConfig, nil
+}
+
+//GetGlobalConfig give config
+func (sa *Adapter) GetGlobalConfig() (*model.GlobalConfig, error) {
+	filter := bson.D{}
+	var result []model.GlobalConfig
+	err := sa.db.globalConfig.Find(filter, &result, nil)
+	if err != nil {
+		return nil, err
+	}
+	if len(result) == 0 {
+		//no record
+		return nil, nil
+	}
+	return &result[0], nil
+
+}
+
+//SaveGlobalConfig saves the global configuration to the storage
+func (sa *Adapter) SaveGlobalConfig(gc *model.GlobalConfig) error {
+	// transaction
+	err := sa.db.dbClient.UseSession(context.Background(), func(sessionContext mongo.SessionContext) error {
+		err := sessionContext.StartTransaction()
+		if err != nil {
+			fmt.Println(err)
+			return err
+		}
+
+		//clear the global config - we always keep only one global config
+		delFilter := bson.D{}
+		_, err = sa.db.globalConfig.DeleteManyWithContext(sessionContext, delFilter, nil)
+		if err != nil {
+			abortTransaction(sessionContext)
+			return err
+		}
+
+		//add the new one
+		_, err = sa.db.globalConfig.InsertOneWithContext(sessionContext, gc)
+		if err != nil {
+			abortTransaction(sessionContext)
+			return err
+		}
+
+		err = sessionContext.CommitTransaction(sessionContext)
+		if err != nil {
+			//TODO print
+			//log.Printf("error on commiting a transaction - %s", err)
+			return err
+		}
+		return nil
+	})
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+//CreateOrganization creates an organization
+func (sa *Adapter) CreateOrganization(name string, requestType string, requiresOwnLogin bool, loginTypes []string, organizationDomains []string) (*model.Organization, error) {
+	now := time.Now()
+
+	orgConfigID, _ := uuid.NewUUID()
+	orgConfig := organizationConfig{ID: orgConfigID.String(), Domains: organizationDomains, DateCreated: now}
+
+	organizationID, _ := uuid.NewUUID()
+	organization := organization{ID: organizationID.String(), Name: name, Type: requestType, RequiresOwnLogin: requiresOwnLogin, LoginTypes: loginTypes,
+		Config: orgConfig, DateCreated: now}
+
+	_, err := sa.db.organizations.InsertOne(organization)
+	if err != nil {
+		return nil, err
+	}
+
+	//return the correct type
+	resOrgConfig := model.OrganizationConfig{ID: orgConfig.ID, Domains: orgConfig.Domains}
+
+	resOrg := model.Organization{ID: organization.ID, Name: organization.Name, Type: organization.Type,
+		RequiresOwnLogin: organization.RequiresOwnLogin, LoginTypes: organization.LoginTypes, Config: resOrgConfig}
+	return &resOrg, nil
+}
+
 //NewStorageAdapter creates a new storage adapter instance
-func NewStorageAdapter(mongoDBAuth string, mongoDBName string, mongoTimeout string) *Adapter {
+func NewStorageAdapter(mongoDBAuth string, mongoDBName string, mongoTimeout string, logger *log.StandardLogger) *Adapter {
 	timeout, err := strconv.Atoi(mongoTimeout)
 	if err != nil {
-		log.Println("Set default timeout - 500")
+		logger.Error("Set default timeout - 500")
 		timeout = 500
 	}
 	timeoutMS := time.Millisecond * time.Duration(timeout)
 
 	db := &database{mongoDBAuth: mongoDBAuth, mongoDBName: mongoDBName, mongoTimeout: timeoutMS}
-	return &Adapter{db: db}
+	return &Adapter{db: db, logger: logger}
+}
+
+func abortTransaction(sessionContext mongo.SessionContext) {
+	err := sessionContext.AbortTransaction(sessionContext)
+	if err != nil {
+		//TODO - log
+	}
+
 }