+++ driver/web/bbsapis.go
@@ -3,20 +3,22 @@ package web
 import (
 	"core-building-block/core"
 	"net/http"
+
+	log "github.com/rokmetro/logging-library/loglib"
 )
 
 //BBsApisHandler handles the APIs implementation used by the platform building blocks
 type BBsApisHandler struct {
-	app *core.Application
+	coreAPIs *core.APIs
 }
 
 //GetTest TODO get test
-func (h BBsApisHandler) GetTest(w http.ResponseWriter, r *http.Request) {
-	res := h.app.BBs.BBsGetTest()
+func (h BBsApisHandler) GetTest(l *log.Log, w http.ResponseWriter, r *http.Request) {
+	res := h.coreAPIs.BBs.BBsGetTest()
 	w.Write([]byte(res))
 }
 
 //NewBBsApisHandler creates new bbs Handler instance
-func NewBBsApisHandler(app *core.Application) BBsApisHandler {
-	return BBsApisHandler{app: app}
+func NewBBsApisHandler(coreAPIs *core.APIs) BBsApisHandler {
+	return BBsApisHandler{coreAPIs: coreAPIs}
 }