+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
@@ -7,8 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Set up unit tests environment. [#7] (https://github.com/rokwire/core-building-block/issues/7)
+- Expose admin API which creates an organization. [#58](https://github.com/rokwire/core-building-block/issues/58)
+- Expose update global config admin API. [#36] (https://github.com/rokwire/core-building-block/issues/36)
+- Expand the model to handle the user devices. [#41] (https://github.com/rokwire/core-building-block/issues/41)
+- Expose get global config admin API [#35] (https://github.com/rokwire/core-building-block/pull/43)
 - Set up auth framework. [#16] (https://github.com/rokwire/core-building-block/issues/16)
 - Expose version API. [#13] (https://github.com/rokwire/core-building-block/issues/13)
-- Set up project skeleton. [#1](https://github.com/rokwire/core-building-block/issues/1)
-- Define data model. [#2](https://github.com/rokwire/core-building-block/issues/2)
-
+- Set up project skeleton. [#1] (https://github.com/rokwire/core-building-block/issues/1)
+- Define data model. [#2] (https://github.com/rokwire/core-building-block/issues/2)
+- Expose create global config admin API [#34] (https://github.com/rokwire/core-building-block/issues/34)
+- Set up logging [#6] (https://github.com/rokwire/core-building-block/issues/6)